Add main loop to game
Add logic to main to setup a new
puzzle and call runGame on it to
loop through our checking functions
for each guess the player makes.

I've also imported some things from
System.IO that we need to get input
from the user. (This was not in the
book.)

From 13.12 Step Three: Making a puzzle,
page 527.